Split os-android.h
Move the shared memory emulation into a new header file in preparation
of merging the Linux and Android operating system header files.
